requirements.txt
# COMS — Cloud Orchestrator and Management System — Requirements
# All free-tier compatible

# Core
streamlit>=1.35.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# LLM (both free)
groq>=0.9.0
google-genai>=1.0.0

# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# Auth (free, local)
streamlit-authenticator>=0.3.3
PyYAML>=6.0

# Data & export
pandas>=2.0.0

# Utilities
requests>=2.31.0
